+++ docs/guides/js/development-execute-bridge-commands-03-read-note.js
@@ -1,6 +1,6 @@
 await runAndLog('read_note', {
   remId: 'qtVwh5XBQbJM2HfSp',
-  includeContent: 'markdown',
+  contentMode: 'markdown',
   depth: 2,
   childLimit: 20,
   maxContentLength: 3000,